rename stacks to porter app
