+++ api/server/handlers/porter_app/create_secret_and_open_pr.go
@@ -1,4 +1,4 @@
-package stacks
+package porter_app
 
 import (
 	"errors"
