+++ api/server/handlers/porter_app/get.go
@@ -1,4 +1,4 @@
-package stacks
+package porter_app
 
 import (
 	"net/http"
