+++ api/server/handlers/porter_app/list_events.go
@@ -1,4 +1,4 @@
-package stacks
+package porter_app
 
 import (
 	"context"
