+++ api/server/handlers/porter_app/parse.go
@@ -1,4 +1,4 @@
-package stacks
+package porter_app
 
 import (
 	"fmt"
